% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{RVersion}
\alias{RVersion}
\alias{$.RVersion}
\title{Return string `"Hello world!"` to R.}
\format{
An object of class \code{environment} of length 1.
}
\usage{
RVersion
}
\description{
Return string `"Hello world!"` to R.
}
\keyword{datasets}
